backend/requirements.txt
pandas==2.3.3
numpy==1.26.4
SQLAlchemy==2.0.46
psycopg2-binary==2.9.11
fastapi==0.115.2
uvicorn==0.30.6
scikit-learn==1.5.2
mlflow==2.14.1
joblib==1.4.2
boto3==1.35.99
langgraph>=0.2.0,<1.0.0
requests==2.32.3
pytest==8.3.5
statsmodels==0.14.4
evidently==0.4.35
imbalanced-learn==0.12.3
xgboost==2.0.3
